Open the hub's west teleport door and survive the west chamber's load
Four findings, each measured through the headless engine harness that can
now load a save state, run frames with scripted input, and read the
dispatch state between them.

The hub's west exit is a teleport door: door type 0x10's attribute becomes
the 0x89 transit family on a west or east door, and the engine then reads
the destination from the header instead of walking to the grid neighbour.
The recovered record had typed it as a plain door, which is why the west
wall either leaked or sealed. The west chamber's east door is the same
type for the way back.

Dungeon_AdjustForTeleportDoors rebuilt the player and camera coordinates
from an 8-bit room id, the one transition helper that still truncated the
bank; it now uses the same 0xff0 row mask the engine's own full-load path
already uses for rooms past 0x100.

The west chamber's header carries layer-effect 8, which the cartridge
defined for itself; the engine's effect dispatch table defines eight
entries and a value past them is a null pointer. Extraction now maps
unknown effects to none until their meaning is measured.

The hub's north door record was a pattern-match false positive: the
top-centre is the staircase, and drawing a door there registered it and
paved the staircase's slot attribute with a door transit strip.

Verified end to end headless: load the hub save, walk left, arrive in the
west chamber, transition settles.

+++ core/zelda3/src/dungeon.h
@@ -396,7 +396,7 @@ void Dungeon_LoadSongBankIfNeeded();
 void Mirror_SaveRoomData();
 void SaveDungeonKeys();
 void Dungeon_AdjustAfterSpiralStairs();
-void Dungeon_AdjustForTeleportDoors(uint8 room, uint8 flag);
+void Dungeon_AdjustForTeleportDoors(uint16 room, uint8 flag);
 void Dungeon_AdjustForRoomLayout();
 void HandleEdgeTransitionMovementEast_RightBy8();
 void Dungeon_StartInterRoomTrans_Right();
